create OneRuleDecoratorRule
