+++ src/Application/Rule/Rules/ConditionRule.php
@@ -10,19 +10,14 @@ use ArtARTs36\MergeRequestLinter\Domain\Rule\RuleDefinition;
 use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\Counter;
 use ArtARTs36\MergeRequestLinter\Shared\Metrics\Value\NullCounter;
 
-class ConditionRule implements RuleDecorator
+class ConditionRule extends OneRuleDecoratorRule
 {
     public function __construct(
-        private Rule $rule,
+        Rule $rule,
         private ConditionOperator $operator,
         private readonly Counter $skippedRules = new NullCounter(),
     ) {
-        //
-    }
-
-    public function getName(): string
-    {
-        return $this->rule->getName();
+        parent::__construct($rule);
     }
 
     public function lint(MergeRequest $request): array
@@ -34,15 +29,5 @@ class ConditionRule implements RuleDecorator
         }
 
         return $this->rule->lint($request);
-    }
-
-    public function getDefinition(): RuleDefinition
-    {
-        return $this->rule->getDefinition();
-    }
-
-    public function getDecoratedRules(): array
-    {
-        return [$this->rule];
     }
 }
