+++ src/Application/Rule/Rules/NonCriticalRule.php
@@ -4,23 +4,9 @@ namespace ArtARTs36\MergeRequestLinter\Application\Rule\Rules;
 
 use ArtARTs36\MergeRequestLinter\Domain\Note\NoteSeverity;
 use ArtARTs36\MergeRequestLinter\Domain\Request\MergeRequest;
-use ArtARTs36\MergeRequestLinter\Domain\Rule\Rule;
-use ArtARTs36\MergeRequestLinter\Domain\Rule\RuleDecorator;
-use ArtARTs36\MergeRequestLinter\Domain\Rule\RuleDefinition;
 
-final class NonCriticalRule implements RuleDecorator
+final class NonCriticalRule extends OneRuleDecoratorRule
 {
-    public function __construct(
-        private readonly Rule $rule,
-    ) {
-        //
-    }
-
-    public function getName(): string
-    {
-        return $this->rule->getName();
-    }
-
     public function lint(MergeRequest $request): array
     {
         $notes = $this->rule->lint($request);
@@ -32,15 +18,5 @@ final class NonCriticalRule implements RuleDecorator
         }
 
         return $notes;
-    }
-
-    public function getDefinition(): RuleDefinition
-    {
-        return $this->rule->getDefinition();
-    }
-
-    public function getDecoratedRules(): array
-    {
-        return [$this->rule];
     }
 }
