+++ tests/Unit/Application/Rule/Rules/NonCriticalRuleTest.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace ArtARTs36\MergeRequestLinter\Tests\Unit\Application\Rule\Rules;
+
+use ArtARTs36\MergeRequestLinter\Tests\TestCase;
+
+final class NonCriticalRuleTest extends TestCase
+{
+    /**
+     * @covers \ArtARTs36\MergeRequestLinter\Application\Rule\Rules\NonCriticalRule::lint
+     */
+    public function testLint(): void
+    {
+
+    }
+}
